Add an option to disable replacement of the Steam binary
LSI was initially created for Solus inclusion, and the best approach
determined was to replace /usr/bin/steam with our own stand in shim,
so that it was fully integrated.

Doing so enables one to run steam from anywhere and it would correctly
appear in the path, with no issues. However, this made LSI suitable
only for distribution integration, such as the very tight integration
seen in Solus.

With this change, the binary name will change to lsi-steam, and an
additional .desktop file to launch the shim will be installed. The
user must then use the LSI specific .desktop file to launch Steam,
at which point they'll be going through the LSI shim system and
reaping the benefits, without having to modify any packages.

More changes are yet to come, which will configure lsi-steam to call
the "real" steam in exec rather than the masked libdir variant.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>