intercept: Teach liblsi-intercept to rewrite dlopen requests for XNA
We currently have a layered block and rewrite system that prevents games
from using vendored SDL, openal, core libs, etc. However the Mono games such
as Stardew Valley will locally load paths using dlopen(), which we've been
blacklisting.

When we now detect a dlopen() to the vendored paths, we'll attempt to rewrite
those to meet the system paths which then forces the dlopen() to resolve to
the system SDL, openAL, etc.

This resolves issue #16.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>