redirect: Start skeleton for new `liblsi-redirect`
We're left in the awkward position that there are some games that **still**
misbehave on Linux, and generally do Dumb Things. So to get around the fact
that some vendors basically stopped caring about us, we'll create a special
`LD_PRELOAD` library that is responsible for redirecting syscalls to basically
monkeypatch broken behaviour completely.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>