common: Add the initial VDF parser
This is a node-tree based parser for the Valve Data Format. Currently
it is provided so that we can parse the Steam Client's library configuration
to determine exactly what directories consitute the "installation" for the
user.

At present, this is a "parse or bail" implementation, and doesn't really
do much in the way of pretty printing errors. It's also not geared for
writing VDF files, though if people request it, I'll happily add that.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>