intercept: Only blacklist common glews (TEMPORARY)
This will permit Super Hexagon to launch, which has a vendored libglew
in tree (1.6 ABI). This is only being allowed on a temporary basis for
two reasons:

 * C ABI and not C++
 * We don't ship this guy in Solus, and don't plan to.

Once LSI is available as a snap, we'll add a super strict mode and then
provide all the GLEW libraries in the image.

This fixes #21.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>