redirect: Add new workaround for Project Highrise
In the current build of Project Highrise, after the first start of the
game settings are stored. Once that leading structure exists, the game
is now able to see files/directories that weren't there on the first run.

The issue however is that instead of loading "prefs/prefs.txt" on Linux,
the game is loading "prefs" directory with fopen64() as if it were a file,
and then attempting to map it. Malloc bombs out and the game instantly
dies.

This is very likely down to an invalid escaping/path join in the game that
could trivially be fixed, however we'll carry this workaround for now until
such point as a new build is out.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>