intercept: Redirect broken x86 dlopen() on 64-bit
This was found while testing Project Highrise, which somehow loads the
plugins from /x86/ and not /x86_64/. To combat that, whilst in 64-bit
intercept mode we'll redirect to the relative ../x86_64 directory if
it also has the same .so file present.

This allows Mono to properly "preload" things once more (such as screen
selector, Steam API, etc.).

Signed-off-by: Ikey Doherty <ikey@solus-project.com>