intercept: When built for snapd, do not override snapd private libs
Certain elements of snapd actually require the (ab)use of snapd to operate,
such as making the NVIDIA proprietary drivers available to the guest snap.
In this instance, if we happen to encounter a library from these search
paths we must unconditionally allow it to be loaded, so that we do not
break libGL resolution.

This adds initial support for issue #30.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>