intercept: Add support for snapd libGL redirection
Under a snapd environment, the dynamic linker gets very confused about
where libGL really is, and this causes 32-bit driver usage under LSI
to completely fall flat.

When snapd support is enabled, we'll check for specific path matches and
attempt to override the input request with the intended bi-arch replacement.
This doesn't yet support multiarch, but that will be added in time.

This is part of the ongoing issue #30.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>