shim: Add new simplified variable management code
This will allow us to extend the number of bootstrap variables we need
to set up for snapd, and actually makes the existing code much simpler
to follow when dealing with existing environmental variables.

With this change we also introduce initial bootstrap for setting the PATH,
as it's incorrect on entering the snap.

This is part of the ongoing issue #30.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>