data: If we're using snap mode, configure frontend to use lsi-exec
This will ensure that the settings UI will always work, which will
somewhat perversely use lsi-exec through the config that it sets to
launch itself to set the config of lsi-exec..

This will allow us to actually expose lsi-settings as a usable thing
within the snap, where it is currently omitted. If we're not using snapd
mode, we'll just keep the file how it used to look.

Signed-off-by: Ikey Doherty <ikey@solus-project.com>